refactor: split lib.rs into focused modules
Extract impl blocks into separate files for better organization:
- column_matching.rs: KeyMatch, CachedColumn, JSON-to-Cell conversion
- config.rs: configure_headers, configure_auth
- request.rs: URL building, path params, HTTP requests
- response.rs: data extraction, pagination handling

lib.rs retains struct definition, Default, singleton, and Guest trait.
No behavior changes — purely organizational restructure.
