+++ src/request.rs
@@ -1,0 +1,408 @@
+//! HTTP request building, URL construction, and API communication
+
+use std::collections::HashMap;
+
+use serde_json::Value as JsonValue;
+
+use crate::bindings::supabase::wrappers::{
+    http, stats, time,
+    types::{Cell, Context, FdwError, FdwResult, Value},
+};
+use crate::spec::OpenApiSpec;
+use crate::{FDW_NAME, OpenApiFdw};
+
+impl OpenApiFdw {
+    /// Fetch and parse the `OpenAPI` spec
+    pub(crate) fn fetch_spec(&mut self) -> Result<(), FdwError> {
+        if let Some(ref url) = self.spec_url {
+            let req = http::Request {
+                method: http::Method::Get,
+                url: url.clone(),
+                headers: self.headers.clone(),
+                body: String::default(),
+            };
+            let resp = http::get(&req)?;
+            http::error_for_status(&resp)
+                .map_err(|err| format!("Failed to fetch OpenAPI spec: {}: {}", err, resp.body))?;
+
+            if resp.body.len() > self.max_response_bytes {
+                return Err(format!(
+                    "OpenAPI spec too large: {} bytes (limit: {} bytes). \
+                     Increase max_response_bytes server option if needed.",
+                    resp.body.len(),
+                    self.max_response_bytes
+                ));
+            }
+
+            let spec_json: JsonValue =
+                serde_json::from_str(&resp.body).map_err(|e| e.to_string())?;
+            let spec = OpenApiSpec::from_json(spec_json)?;
+
+            // Use base_url from spec if not explicitly set
+            if self.base_url.is_empty() {
+                if let Some(url) = spec.base_url() {
+                    self.base_url = url.trim_end_matches('/').to_string();
+                }
+            }
+
+            self.spec = Some(spec);
+            stats::inc_stats(FDW_NAME, stats::Metric::BytesIn, resp.body.len() as i64);
+        }
+        Ok(())
+    }
+
+    /// Extract a qual value as a string
+    pub(crate) fn qual_value_to_string(
+        qual: &crate::bindings::supabase::wrappers::types::Qual,
+    ) -> Option<String> {
+        if qual.operator() != "=" {
+            return None;
+        }
+        if let Value::Cell(cell) = qual.value() {
+            match cell {
+                Cell::String(s) => Some(s),
+                Cell::I32(n) => Some(n.to_string()),
+                Cell::I64(n) => Some(n.to_string()),
+                Cell::F32(n) => Some(n.to_string()),
+                Cell::F64(n) => Some(n.to_string()),
+                Cell::Bool(b) => Some(b.to_string()),
+                _ => None,
+            }
+        } else {
+            None
+        }
+    }
+
+    /// Resolve a relative or absolute pagination URL against the base URL and endpoint.
+    ///
+    /// Handles four forms of `next_url`:
+    /// - Absolute URLs (`http://...`, `https://...`) → used as-is
+    /// - Query-only (`?page=2`) → resolves against `base_url + endpoint`
+    /// - Absolute paths (`/items?page=2`) → resolves against `base_url`
+    /// - Bare relative paths (`page/2`) → resolves against `base_url/`
+    pub(crate) fn resolve_pagination_url(&self, next_url: &str) -> String {
+        if next_url.starts_with("http://") || next_url.starts_with("https://") {
+            next_url.to_string()
+        } else if next_url.starts_with('?') {
+            let endpoint_base = self.endpoint.split('?').next().unwrap_or(&self.endpoint);
+            format!("{}{endpoint_base}{next_url}", self.base_url)
+        } else if next_url.starts_with('/') {
+            format!("{}{next_url}", self.base_url)
+        } else {
+            format!("{}/{next_url}", self.base_url)
+        }
+    }
+
+    /// Substitute path parameters in endpoint template from quals.
+    ///
+    /// Writes substituted values into `injected` so they can be re-injected
+    /// into result rows (ensuring PostgreSQL's post-filter passes).
+    ///
+    /// Returns (resolved_endpoint, path_params_used) where path_params_used
+    /// contains lowercase names of parameters that were substituted.
+    ///
+    /// # Errors
+    /// Returns an error if required path parameters are missing from quals.
+    pub(crate) fn substitute_path_params(
+        endpoint: &str,
+        quals: &[crate::bindings::supabase::wrappers::types::Qual],
+        injected: &mut HashMap<String, String>,
+    ) -> Result<(String, Vec<String>), String> {
+        if !endpoint.contains('{') {
+            return Ok((endpoint.to_string(), Vec::new()));
+        }
+
+        // Build a map of qual field -> value for path parameter substitution
+        // Pre-allocate for 2 entries per qual (original + lowercase key)
+        let mut qual_map: HashMap<String, String> = HashMap::with_capacity(quals.len() * 2);
+        for qual in quals {
+            if let Some(value) = Self::qual_value_to_string(qual) {
+                // Store both original and lowercase versions for flexible matching
+                qual_map.insert(qual.field().to_lowercase(), value.clone());
+                qual_map.insert(qual.field(), value);
+            }
+        }
+
+        let mut endpoint = endpoint.to_string();
+        let mut path_params_used: Vec<String> = Vec::new();
+        let mut missing_params: Vec<String> = Vec::new();
+
+        // Find all {param} patterns and substitute
+        while let Some(start) = endpoint.find('{') {
+            if let Some(end) = endpoint[start..].find('}') {
+                let param_name = &endpoint[start + 1..start + end];
+                let param_lower = param_name.to_lowercase();
+
+                // Try to find matching qual (case-insensitive)
+                let value = qual_map
+                    .get(&param_lower)
+                    .or_else(|| qual_map.get(param_name));
+
+                if let Some(val) = value {
+                    path_params_used.push(param_lower.clone());
+                    // Store the path param for injection into rows (unencoded for PostgreSQL filter)
+                    injected.insert(param_lower, val.clone());
+                    endpoint = format!(
+                        "{}{}{}",
+                        &endpoint[..start],
+                        urlencoding::encode(val),
+                        &endpoint[start + end + 1..]
+                    );
+                } else {
+                    // Track missing parameter and remove it from the endpoint to continue
+                    missing_params.push(param_name.to_string());
+                    endpoint = format!("{}{}", &endpoint[..start], &endpoint[start + end + 1..]);
+                }
+            } else {
+                break;
+            }
+        }
+
+        // Return error if any required path parameters are missing
+        if !missing_params.is_empty() {
+            return Err(format!(
+                "Missing required path parameter(s) in WHERE clause: {}. \
+                 Add WHERE {} to your query.",
+                missing_params.join(", "),
+                missing_params
+                    .iter()
+                    .map(|p| format!("{p} = '<value>'"))
+                    .collect::<Vec<_>>()
+                    .join(" AND ")
+            ));
+        }
+
+        Ok((endpoint, path_params_used))
+    }
+
+    /// Build query parameters from pagination state, quals, and API key.
+    ///
+    /// Returns (url_params, injected_entries) where injected_entries are
+    /// qual values to merge into `self.injected_params` for row injection.
+    /// Excludes path parameters and rowid column.
+    pub(crate) fn build_query_params(
+        &self,
+        quals: &[crate::bindings::supabase::wrappers::types::Qual],
+        path_params_used: &[String],
+    ) -> (Vec<String>, Vec<(String, String)>) {
+        // Pre-allocate for cursor + page_size + quals + api_key
+        let mut params = Vec::with_capacity(quals.len() + 3);
+        let mut injected_entries = Vec::new();
+
+        // Add pagination cursor if we have one
+        if let Some(ref cursor) = self.next_cursor {
+            params.push(format!(
+                "{}={}",
+                self.cursor_param,
+                urlencoding::encode(cursor)
+            ));
+        }
+
+        // Add page size if configured
+        if self.page_size > 0 && !self.page_size_param.is_empty() {
+            params.push(format!("{}={}", self.page_size_param, self.page_size));
+        }
+
+        // Add remaining quals as query params (exclude path params and rowid)
+        for qual in quals {
+            let field_lower = qual.field().to_lowercase();
+
+            // Skip if used as path param
+            if path_params_used.contains(&field_lower) {
+                continue;
+            }
+
+            // Skip the rowid column
+            if field_lower == self.rowid_col {
+                continue;
+            }
+
+            if let Some(value) = Self::qual_value_to_string(qual) {
+                // Track for injection back into rows
+                // (so PostgreSQL's WHERE filter passes even if the API doesn't echo it back)
+                injected_entries.push((field_lower, value.clone()));
+                params.push(format!(
+                    "{}={}",
+                    urlencoding::encode(&qual.field()),
+                    urlencoding::encode(&value)
+                ));
+            }
+        }
+
+        // Add API key as query parameter if configured
+        if let Some((ref param_name, ref param_value)) = self.api_key_query {
+            params.push(format!(
+                "{}={}",
+                urlencoding::encode(param_name),
+                urlencoding::encode(param_value)
+            ));
+        }
+
+        (params, injected_entries)
+    }
+
+    /// Build the URL for a request, handling path parameters and pagination.
+    ///
+    /// Updates `self.injected_params` in place (avoids cloning on pagination).
+    ///
+    /// Supports endpoint templates like:
+    /// - `/users/{user_id}/posts`
+    /// - `/projects/{org}/{repo}/issues`
+    /// - `/resources/{type}/{id}`
+    ///
+    /// Path parameters are substituted from WHERE clause quals.
+    ///
+    /// # Errors
+    /// Returns an error if required path parameters are missing from the WHERE clause.
+    pub(crate) fn build_url(&mut self, ctx: &Context) -> Result<String, String> {
+        // Use next_url for pagination if available (injected_params unchanged)
+        if let Some(ref next_url) = self.next_url {
+            return Ok(self.resolve_pagination_url(next_url));
+        }
+
+        let quals = ctx.get_quals();
+
+        // Substitute path parameters (no self borrow — takes &mut injected_params directly)
+        let (endpoint, path_params_used) =
+            Self::substitute_path_params(&self.endpoint, &quals, &mut self.injected_params)?;
+
+        // Check for rowid pushdown for single-resource access
+        // Only if endpoint doesn't already have path params and rowid qual exists
+        if path_params_used.is_empty() {
+            if let Some(id_qual) = quals
+                .iter()
+                .find(|q| q.field().to_lowercase() == self.rowid_col && q.operator() == "=")
+            {
+                if let Some(id) = Self::qual_value_to_string(id_qual) {
+                    self.injected_params
+                        .insert(self.rowid_col.clone(), id.clone());
+                    return Ok(format!(
+                        "{}{}/{}",
+                        self.base_url,
+                        endpoint,
+                        urlencoding::encode(&id)
+                    ));
+                }
+            }
+        }
+
+        // Build query parameters
+        let (params, injected_entries) = self.build_query_params(&quals, &path_params_used);
+        self.injected_params.extend(injected_entries);
+
+        // Assemble final URL
+        let mut url = format!("{}{}", self.base_url, endpoint);
+        if !params.is_empty() {
+            let separator = if url.contains('?') { '&' } else { '?' };
+            url.push(separator);
+            url.push_str(&params.join("&"));
+        }
+
+        Ok(url)
+    }
+
+    /// Make a request to the API with automatic rate limit handling
+    pub(crate) fn make_request(&mut self, ctx: &Context) -> FdwResult {
+        let url = self.build_url(ctx)?;
+
+        let req = http::Request {
+            method: self.method,
+            url,
+            headers: self.headers.clone(),
+            body: self.request_body.clone(),
+        };
+
+        // Retry loop for rate limiting (HTTP 429)
+        let mut retry_count = 0;
+        const MAX_RETRIES: u32 = 3;
+        const MAX_RETRY_DELAY_MS: u64 = 30_000;
+
+        let resp = loop {
+            let resp = match req.method {
+                http::Method::Post => http::post(&req)?,
+                _ => http::get(&req)?,
+            };
+
+            // Handle rate limiting (HTTP 429)
+            if resp.status_code == 429 {
+                if retry_count >= MAX_RETRIES {
+                    return Err(format!(
+                        "API rate limit exceeded after {MAX_RETRIES} retries. \
+                         Consider adding a page_size option to reduce request frequency."
+                    ));
+                }
+
+                // Try to get retry delay from Retry-After header (case-insensitive),
+                // capped to prevent absurdly long waits from malicious/buggy servers
+                let delay_ms = resp
+                    .headers
+                    .iter()
+                    .find(|h| h.0.to_lowercase() == "retry-after")
+                    .and_then(|h| h.1.parse::<u64>().ok())
+                    .map(|secs| secs.saturating_mul(1000).min(MAX_RETRY_DELAY_MS))
+                    .unwrap_or_else(|| {
+                        // Exponential backoff: 1s, 2s, 4s (capped)
+                        let backoff = 1000u64.saturating_mul(1 << retry_count);
+                        backoff.min(MAX_RETRY_DELAY_MS)
+                    });
+
+                time::sleep(delay_ms);
+                retry_count += 1;
+                continue;
+            }
+
+            break resp;
+        };
+
+        if self.debug_timing {
+            crate::bindings::supabase::wrappers::utils::report_info(&format!(
+                "[openapi_fdw] HTTP {} {} -> {} ({} bytes)",
+                if matches!(req.method, http::Method::Post) {
+                    "POST"
+                } else {
+                    "GET"
+                },
+                req.url,
+                resp.status_code,
+                resp.body.len()
+            ));
+        }
+
+        // Handle 404 as empty result (no matching resource)
+        if resp.status_code == 404 {
+            self.src_rows = Vec::new();
+            self.src_idx = 0;
+            self.next_cursor = None;
+            self.next_url = None;
+            return Ok(());
+        }
+
+        http::error_for_status(&resp).map_err(|err| format!("{}: {}", err, resp.body))?;
+
+        if resp.body.len() > self.max_response_bytes {
+            return Err(format!(
+                "Response body too large: {} bytes (limit: {} bytes). \
+                 Increase max_response_bytes server option if needed.",
+                resp.body.len(),
+                self.max_response_bytes
+            ));
+        }
+
+        let mut resp_json: JsonValue =
+            serde_json::from_str(&resp.body).map_err(|e| e.to_string())?;
+
+        stats::inc_stats(FDW_NAME, stats::Metric::BytesIn, resp.body.len() as i64);
+
+        // Handle pagination before extracting data (borrows resp_json)
+        self.handle_pagination(&resp_json);
+
+        // Extract data by taking ownership (avoids cloning the array)
+        self.src_rows = self.extract_data(&mut resp_json)?;
+        self.src_idx = 0;
+
+        // Build column key map for O(1) lookups during iter_scan
+        self.build_column_key_map();
+
+        Ok(())
+    }
+}
